Lab 1 III & IV: schedule map and reduce tasks
Start a waitGroup for all the tasks.

For each map/reduce task, create a go routine to finish the task. In the
go routine, wait for an available worker from the register channel, and
call its RPC to do the actual map/reduce. It then waits for the RPC call
to return;

If it succeeds, call waitGroup.Done() for the waitGroup and pub the
worker back to the register channel so it can be used by other tasks.
If it fails, give up that worker and wait for a new worker.
(The failed worker won't be put back in register channel every again)

Note, we call waitGroup.Done() before put worker back because the
register channel has cap=0, so for the last few tasks, the worker can't
be put back because there's no more task and the channel will block.
So in effect, the last few assignTask go routines never return, but the
main routine will finish once the waitGroup is all done.