Lab 3A: key/value service without log compaction
Fixed a bug in raft.

Some notes:

To do correct Get, the server must be in a majority of the cluster so
that it’s not a partitioned leader with stale data.

Consider: leader just did a Put, and returned to client successfully,
then fail. The client then Get the same key, and got notified the server
is not leader anymore, so client find the leader and wants to do Get.

The new leader might not have applied the Put op (although it must have
that Put in the log). The new Get will put a Get entry in the log that’s
after the Put. Here we must not get the value for the Get right way,
but have to wait till the Get entry is applied so that it can get the
latest k/v, after the Put is applied.

On the other hand, if client do a Get, the server successfully commit
the Get, but fail to return the Get value to client, then server fail.
Client will find new leader, the new leader haven’t applied this Get,
and client will retry this Get and add a new entry in the log. The retry
Get RPC will wait till the old Get entry is applied and returned, and
the new Get entry created by the retry can be ignored when applied.